requirements.txt
Django==5.0.1
django-ninja==1.1.0
psycopg2-binary==2.9.9
pyjwt==2.8.0